requirements.txt
fastapi
uvicorn
sqlalchemy>=2
pydantic
python-dotenv
cryptography
rfc8785
eth-hash
requests

